+++ docs/realm.js
@@ -456,7 +456,7 @@ class Realm {
  * @property {boolean} [indexed] - Signals if this property should be indexed. Only supported for
  *   `"string"`, `"int"`, and `"bool"` properties.
  * @property {string} [alias] - an alias in JavaScript for the underlying name in the Realm file. If set, this must be
- *   used when reading/writing the property or making queries against it. Trying to use the internal name will throw
+ *   used when reading/writing the property or making queries against it. Trying to use the underlying name will throw
  *   an exception.
  */